adding common methods
