+++ src/main/java/com/blackducksoftware/integration/hub/api/view/ImprovedCommonNotificationState.java
@@ -1,0 +1,10 @@
+package com.blackducksoftware.integration.hub.api.view;
+
+import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
+
+public interface ImprovedCommonNotificationState extends ReducedCommonNotificationState {
+    public NotificationContent getNotificationContent();
+
+    public void setNotificationContent(NotificationContent notificationContent);
+
+}
