+++ src/main/java/com/blackducksoftware/integration/hub/api/view/ImprovedNotificationView.java
@@ -23,13 +23,10 @@
  */
 package com.blackducksoftware.integration.hub.api.view;
 
-import java.util.Date;
-
-import com.blackducksoftware.integration.hub.api.generated.enumeration.NotificationType;
 import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
 
 // This is named ImprovedNotificationView because it adds the 'content' member with an actual type.
-public class ImprovedNotificationView extends ReducedNotificationView implements CommonNotificationState {
+public class ImprovedNotificationView extends ReducedNotificationView implements ImprovedCommonNotificationState {
     public NotificationContent notificationContent;
 
     @Override
@@ -38,48 +35,8 @@ public class ImprovedNotificationView extends ReducedNotificationView implements
     }
 
     @Override
-    public String getContentType() {
-        return contentType;
-    }
-
-    @Override
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    @Override
-    public NotificationType getType() {
-        return type;
-    }
-
-    @Override
-    public String getJson() {
-        return json;
-    }
-
-    @Override
     public void setNotificationContent(final NotificationContent notificationContent) {
         this.notificationContent = notificationContent;
-    }
-
-    @Override
-    public void setContentType(final String contentType) {
-        this.contentType = contentType;
-    }
-
-    @Override
-    public void setCreatedAt(final Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    @Override
-    public void setType(final NotificationType type) {
-        this.type = type;
-    }
-
-    @Override
-    public void setJson(final String json) {
-        this.json = json;
     }
 
 }
