+++ src/main/java/com/blackducksoftware/integration/hub/api/view/ReducedCommonNotificationState.java
@@ -1,11 +1,8 @@
 package com.blackducksoftware.integration.hub.api.view;
 
 import com.blackducksoftware.integration.hub.api.generated.enumeration.NotificationType;
-import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
 
-public interface CommonNotificationState {
-    public NotificationContent getNotificationContent();
-
+public interface ReducedCommonNotificationState {
     public String getContentType();
 
     public java.util.Date getCreatedAt();
@@ -13,8 +10,6 @@ public interface CommonNotificationState {
     public NotificationType getType();
 
     public String getJson();
-
-    public void setNotificationContent(NotificationContent notificationContent);
 
     public void setContentType(String contentType);
